# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Внутренний API дневника тренировок",
    "description": "API для доверенных сервисов. X-Api-Key проверяет вызывающий сервис; X-Telegram-User-Id выбирает пользователя. Не передавайте ключ клиентским приложениям: он позволяет действовать от имени любого пользователя. Вес в кг, расстояние в км, длительность в секундах.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "apiKey": [],
      "telegramIdentity": []
    }
  ],
  "paths": {
...
    "/api/training/history": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/history": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "history_1",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/progress": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "progress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "quarter"
            }
          },
          {
            "name": "workingOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "bySession",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Progress"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{id}/history": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "exerciseHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/{period}": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "report",
        "parameters": [
          {
            "name": "period",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workingOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "equalElapsed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SetPlan": {
        "type": "object",
        "properties": {
          "reps": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "distance": {
            "type": "number"
          },
          "warmup": {
            "type": "boolean"
          }
        }
      },
...
      "SetInput": {
        "type": "object",
        "properties": {
          "sessionExerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "weight": {
            "type": "number"
          },
          "repetitions": {
            "type": "integer",
            "format": "int32"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "distance": {
            "type": "number"
          },
          "rpe": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "warmup": {
            "type": "boolean"
          }
        }
      },
      "ExerciseView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "metricType": {
            "type": "string",
            "enum": [
              "STRENGTH",
              "BODYWEIGHT",
              "CARDIO",
              "TIMED"
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "targetSets": {
            "type": "integer",
            "format": "int32"
          },
          "targetReps": {
            "type": "integer",
            "format": "int32"
          },
          "targetWeight": {
            "type": "number"
          },
          "restSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetView"
            }
          },
          "targetDuration": {
            "type": "integer",
            "format": "int32"
          },
          "targetDistance": {
            "type": "number"
          },
          "plan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetPlan"
            }
          },
          "skipped": {
            "type": "boolean"
          }
        }
      },
      "SessionView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "templateId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentPosition": {
            "type": "integer",
            "format": "int32"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseView"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "pausedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pausedSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "restUntil": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SetView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number"
          },
          "repetitions": {
            "type": "integer",
            "format": "int32"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "distance": {
            "type": "number"
          },
          "rpe": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "warmup": {
            "type": "boolean"
          }
        }
      },
      "Summary": {
        "type": "object",
        "properties": {
          "durationSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "exercises": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "repetitions": {
            "type": "integer",
            "format": "int64"
          },
          "volume": {
            "type": "number"
          }
        }
      },
...
      "HistoricalExercise": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetInput"
            }
          }
        }
      },
      "HistoricalInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoricalExercise"
            }
          },
          "pausedSeconds": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Point": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        }
      },
      "Progress": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "metric": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Point"
            }
          },
          "changePercent": {
            "type": "number"
          },
          "records": {
            "$ref": "#/components/schemas/Records"
          }
        }
      },
      "Records": {
        "type": "object",
        "properties": {
          "maxWeight": {
            "type": "number"
          },
          "maxReps": {
            "type": "integer",
            "format": "int32"
          },
          "bestSetVolume": {
            "type": "number"
          },
          "bestSessionVolume": {
            "type": "number"
          },
          "estimatedOneRm": {
            "type": "number"
          }
        }
      },
      "Day": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "workouts": {
            "type": "integer",
            "format": "int64"
          },
          "volume": {
            "type": "number"
          }
        }
      },
      "Report": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date"
          },
          "end": {
            "type": "string",
            "format": "date"
          },
          "current": {
            "$ref": "#/components/schemas/Totals"
          },
          "previous": {
            "$ref": "#/components/schemas/Totals"
          },
          "workoutChange": {
            "type": "integer",
            "format": "int64"
          },
          "volumeChangePercent": {
            "type": "number"
          },
          "frequency": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Day"
            }
          }
        }
      },
      "Totals": {
        "type": "object",
        "properties": {
          "workouts": {
            "type": "integer",
            "format": "int64"
          },
          "exercises": {
            "type": "integer",
            "format": "int64"
          },
          "sets": {
            "type": "integer",
            "format": "int64"
          },
          "repetitions": {
            "type": "integer",
            "format": "int64"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "volume": {
            "type": "number"
          }
        }
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header"
      },
      "telegramIdentity": {
        "type": "apiKey",
        "name": "X-Telegram-User-Id",
        "in": "header"
      }
    }
  }
}